pyproject.toml
[tool.poetry]
name = "artrade"
version = "0.0.1"
description = "A standalone simple trading system for retail trader"
authors = ["chrischen <chrischen@ignity.xyz>"]
maintainers = ["chrischen <chrischen@ignity.xyz>"]
license = "MIT"
readme = "README.org"
repository = "https://github.com/chrischen3121/aurora-trade"
keywords = ["trading", "quantitative", "finance"]

classifiers = [
    "Development Status :: 1 - Planning",
    "Framework :: AsyncIO",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: C++",
]

[tool.poetry.dependencies]
python = "^3.13"
longport = "^2.1.5"
pydantic = "^2.10.4"
cryptography = "^44.0.0"
platformdirs = "^4.3.6"
python-dotenv = "^1.0.1"
loguru = "^0.7.3"
protobuf = "^5.29.2"

[tool.poetry.group.dev.dependencies]
ipython = "^8.31.0"
sphinx = "^8.1.3"
pytest = "^8.3.4"
pympler = "^1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
